Poly-commitment: rename evaluation_proof into ipa
Result of:
```
find arrabiata -name "*.rs" -type f -exec sed -i 's/evaluation_proof/ipa/g' {} \;
```
(replace arrabiata with folding, ivc, kimchi, msg, o1vm, poly_commitment, tools)

followed by cargo +nightly fmt and `git mv poly-commitment/src/evaluation_proof.rs poly-commitment/src/ipa.rs`.
