+++ kimchi/src/snarky/folding.rs
@@ -8,7 +8,7 @@ use crate::{
 use ark_ec::AffineRepr;
 use ark_ff::{BigInteger, One, PrimeField};
 use mina_curves::pasta::Fp;
-use poly_commitment::{evaluation_proof::OpeningProof, OpenProof};
+use poly_commitment::{ipa::OpeningProof, OpenProof};
 use std::{marker::PhantomData, ops::Add};
 
 mod instance;
