+++ kimchi/src/tests/xor.rs
@@ -22,7 +22,7 @@ use mina_poseidon::{
 use num_bigint::BigUint;
 use o1_utils::{BigUintHelpers, BitwiseOps, FieldHelpers, RandomField};
 use poly_commitment::{
-    evaluation_proof::{endos, OpeningProof, SRS},
+    ipa::{endos, OpeningProof, SRS},
     SRS as _,
 };
