+++ poly-commitment/tests/commitment.rs
@@ -12,7 +12,7 @@ use poly_commitment::{
         combined_inner_product, BatchEvaluationProof, BlindedCommitment, CommitmentCurve,
         Evaluation, PolyComm,
     },
-    evaluation_proof::{DensePolynomialOrEvaluations, OpeningProof, SRS},
+    ipa::{DensePolynomialOrEvaluations, OpeningProof, SRS},
     SRS as _,
 };
 use rand::{CryptoRng, Rng, SeedableRng};
